fix: reorganize stack subcharts to fix the enable toggle for subcharts (#13)
Helm seems to choke on condition statements in subcharts that themselves have dependencies.
Managing the first-level subcharts as dependencies via Chart.yaml (rather than including
them directly in the charts/ directory) seems to resolve the issue. In local testing,
this fixes the problem, and when this PR is merged I will confirm it still works as
expected for installing from the remote repository.

This new layout is similar to what we tried previously, except that the subcharts of stack
are in an "internal" directory rather than the top level, to signify that they are not
useable in isolation.