feat: use observe endpoint to auto-configure scheme/host/port/customerID
Rather than generating collector URLs using a provided scheme, customer,
host, and port, also allow the entire address (minus the URL path, which
changes per agent) as a single value. If this value is provided, it
overrides the others. This change also adds enforcement that if the
full address is not provided, then the customer ID must be provided
or the template generation will fail with a descriptive error.