Add `host_filter: "true"` to default metrics values. Chart renders with an empty value without this.
Further in the values, the configmap passed to grafana references a `Values.prom_config.host_filter` which is never initialized.

In our environment this was resulting in memory spikes for the grafana agent that cause an OOM crash loop. Adding this value in as a default appears to stop this, and brings things in line with the previously-used kustomize definition.