fix: remove configmaps from tests
Each kubernetes manifest is treated as a standalone test by helm,
so creating configmaps to configure test pods does not work as
expected -- configmaps either persist, which is a bug, or they
are cleaned up before the job relying on them can be run.
Avoid them entirely for the configuration of test pods.