feat: cleanup tagging for (mostly) metrics
1. in k8sattributes, stop adding all labels and annotations as tags.
As that is a very pretty big list with unclear value. We should enable
them selectively when user needs them.

2. removed k8sattributes/podcontroller, no one is using them

3. dropped resourcedetection for all deployments
This is the one that tags a lot of node level information. But, for
deployments, it tags about the node that the deployment runs on, instead
of the node that runs the things being monitored (e.g. pods). So, the
result is very confusing. Let's just drop it. We can add selective tags
back when they are useful.

4. change the debug_objectSource tag to debug_source
Since it was initially used for entity/object streams only. But now it's
being used for logs & metrics data too.