feat: Add facets for several new resources
Add facets for:

 - ConfigMap
 - Secret
 - PersistentVolume
 - PersistentVolumeClaim
 - Service
 - Ingress
 - RoleBinding
 - ClusterRoleBinding
 - ServiceAccount

Add name and uid for all entities (including already present ones).

These will be used by frontend to add correlation tags and enable
pivoting between pages.

Remove CronJob "active" facet
This facet was computed using OTTL's Len() on a slice of pointers, which
is not a supported ValueType for Len().

Therefore, this facet will have to be computed with a custom processor.