chore: Remove helm-docs footer
On one hand, it is not possible to generate the helm-docs version footer when
helm-docs is installed via go install, which is the most common and
easiest way to install this tool.

On the other hand, configuring helm-docs (when installed via brew or
other package managers) to skip printing its version footer is very
easy (passing --skip-version-footer).

Make skipping the version footer the default behavior to facilitate
development when helm-docs is installed via go install.