feat: Cleanup values.yaml with sensible defaults (#256)
Send everything to the logs path ("/v2/otel").
This means that we can remove the entity-specific exporter, and use
either the ENTITY or OBSERVE token as authorization in the logs path.

Remove image tag and pull policy from default values: this means that
the helm charts, unless we override these with custom values, will use
the latest (publicly released) agent image, and will ALWAYS pull it.

If we specify a custom image, imagePullPolicy is automatically changed
to "IfNotPresent", which is what we want. To avoid confusion, since the
pull policy has amazing defaults, we completely remove it!

Change deployment template: remove chartVersion as default image tag,
since the two are completely uncorrelated and it makes no sense to use
that default. Instead, only specify an image tag when it's set
explicitly in the variables. Otherwise, just omit the tag, which
defaults to "latest".

Same goes for the iamgePullPolicy, only set it when explicitly defined,
otherwise use the defaults!