fix: Unify pull/watch events before observek8sattributes
The unification of events is currently defined in transform/object,
which was moved after observek8sattributes due to the Secret redaction
requirements.

However, the observek8sattributes custom processor expects the data from
pull/watch to be unified.

The solution to this is to extract the unification logic into a
different processor, which can be injected in the pipeline before
observek8sattributes.