fix: Remove container information from pod metrics
Container information currently is not reliable, a random
container id / name is being discovered, which can lead
to confusion.

It should be possible to uniquely identify the container
based on the pod, but I couldn't get this to work. That
would still be the ideal solution - we can always add this
back if we manage to get that to work.

Fixes OB-45911