fix: Treat "nil" ready replicas as 0 (#420)
We currently set the status of various workload objects to "Unhealthy"
when readyReplicas != desiredReplicas. However, sometimes readyReplicas
could be omitted (instead of being set to 0) if no replicas are ready.

At the same time, in particular cases, one could set desiredReplicas to
0 (for whatever reason).

In this case, with the current logic, the object would always be
considered "Unhealthy" since nil != 0.

Treat nil as 0 to consider objects defined with desiredReplicas == 0
always "Healthy".