feat: OB-41415 - add logging for fargate
In a previous PR, we added metric collection from Fargate. To do log collection, we rely more on
the customer. We don't have access to the underlying node, so we cannot collect logs directly
from there. Instead, we ask the customer to send their logs to an emptyDir volume that our otel
collector sidecar can read from.