fix: OB-52735 - pin otelcollector in operator to a version to prevent breaking changes from propagating without warning
We saw an issue where a breaking change in 0.142.0 caused the collector that we use for the operator (for EKS Fargate) to error. A flag changed from default off to on, causing the resourcedetection/cloud processor (and therefore the collector) to fail. We should version here, instead of relying on latest, to prevent this from happening again. We will work on the underlying bug (an error with the eks detector) after we get functionality back.